Determine unique values from data, not pandas.cat
When using the meta-data of the pandas categorical dtype to determine
all unique values, it might create an encoding for values not present in
the data. Ignoring the meta-data and using the seen unique values
instead ensures proper behavior when a new value occurs in the test
data. It will be handled through the handle_unknown strategy
accordingly.