Test unknown value encoded correctly for pd cat
Unknown values should be determined based on the training data, not
the pandas categorical meta-data. This new test checks that the value
of 'los angeles' that is specified in the series' dtype is still
encoded as an unknown value in the test data, because it was not present
in the training data. See also #235.