In HashEncoder use shared memory instead of queue for multiproccessing
It is faster to write directly in memory that to have to data transit
through a queue.

The multiprocessing method is similar to what it was with queues: the
dataframe is split into chunks, and each process applies the hashing
trick to its chunk of the dataframe. Instead of writting the result
to a queue, it writes it directly in a shared memory segment, that is
also the underlying memory of a numpy array that is used to build the
output dataframe.

Tested on MacOS Monterey, 2.3 GHz Quad-Core i7, it is between 1.3 and
1.6 times faster depending on the df size and how many processes are
used.