Allow forking processes instead of spwaning them and make it default
This makes the HashEncoder transform method a lot faster on small
datasets.

The spawn process creation method creates a new python interpreter
from scratch, and re-import all required module. In a minimal case
(pandas and category_encoders.hashing only are imported) this adds
a ~2s overhead to any call to transform.

Fork creates a copy of the current process, and that's it. It is
unsafe to use with threads, locks, file descriptors, ... but in that
case the only thing the forked process will do is process some data
and write it to ITS OWN segment of a shared memory. It is a lot
faster as pandas doesn't have to be re-imported (around 20ms?)

It might take up more memory as more than the necessary variables
(the largest one by far being the HashEncoder instance, which
include the user dataframe) will be copied. Add the option to use
spawn instead of fork to potentially save some memory.